refactor!: remove mono repo, delete modal, create monopackage
