+++ .gitignore
@@ -3,6 +3,7 @@ node_modules
 site/
 
 dist/
+.eik
 
 *.log
 *.swp
