--- lerna.json
@@ -1,6 +1,0 @@
-{
-  "npmClient": "yarn",
-  "registry": "https://registry.npmjs.org",
-  "useWorkspaces": true,
-  "version": "independent"
-}
