+++ package.json
@@ -3,12 +3,15 @@
   "version": "0.0.0",
   "private": true,
   "description": "Custom elements for Fabric",
-  "workspaces": [
-    "packages/*"
+  "exports": "./dist/index.js",
+  "module": "dist/index.js",
+  "files": [
+    "dist/"
   ],
+  "types": "dist/index.d.ts",
   "scripts": {
     "prebuild": "rimraf packages/*/dist",
-    "build": "NODE_ENV=production lerna run build --stream",
+    "build": "NODE_ENV=production microbundle -i index.js -o dist/index.js --no-pkg-main --format modern",
     "format": "prettier --write . --ignore-path .gitignore",
     "lint": "npm run lint:format && npm run lint:eslint",
     "lint:format": "prettier --check . --ignore-path .gitignore",
@@ -39,12 +42,17 @@
     "rimraf": "^3.0.2",
     "tailwindcss": "^2.0.2",
     "vite": "^2.0.0-beta.56",
-    "vite-plugin-html": "^2.0.0-beta.6"
+    "vite-plugin-html": "^2.0.0-beta.6",
+    "lit-element": "^2.4.0",
+    "lit-html": "^1.3.0",
+    "@chbphone55/classnames": "^2.0.0",
+    "@fabric-ds/component-classes": "^0.0.31",
+    "element-collapse": "^0.9.1",
+    "microbundle": "^0.13.3",
+    "typescript": "^4.3.5"
   },
   "dependencies": {
-    "html-format": "^1.0.1",
-    "lit-element": "^2.4.0",
-    "lit-html": "^1.3.0"
+    "html-format": "^1.0.1"
   },
   "publishConfig": {
     "access": "public"
