--- packages/box/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "@fabric-ds/elements-box",
-  "version": "0.0.3",
-  "description": "Custom elements for fabric",
-  "type": "module",
-  "exports": "./dist/fabric-box.js",
-  "module": "dist/fabric-box.js",
-  "files": [
-    "dist/"
-  ],
-  "types": "dist/fabric-box.d.ts",
-  "scripts": {
-    "build": "yarn build:js",
-    "build:js": "esbuild src/fabric-box.js --format=esm --outfile=dist/fabric-box.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/fabric-ds/elements"
-  },
-  "license": "ISC",
-  "devDependencies": {
-    "@fabric-ds/component-classes": "^0.0.31"
-  },
-  "dependencies": {
-    "lit-element": "^2.4.0",
-    "lit-html": "^1.3.0"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "gitHead": "a8154fa0605e4c7e3fa02f6058b10e4c38b0471b"
-}
