--- packages/breadcrumbs/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@fabric-ds/elements-breadcrumbs",
-  "version": "0.0.3",
-  "description": "Custom elements for fabric",
-  "type": "module",
-  "exports": "./dist/fabric-breadcrumbs.js",
-  "module": "dist/fabric-breadcrumbs.js",
-  "files": [
-    "dist/"
-  ],
-  "types": "dist/fabric-breadcrumbs.d.ts",
-  "scripts": {
-    "build": "yarn build:js",
-    "build:js": "esbuild src/fabric-breadcrumbs.js --format=esm --outfile=dist/fabric-breadcrumbs.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/fabric-ds/elements"
-  },
-  "license": "ISC",
-  "gitHead": "a8154fa0605e4c7e3fa02f6058b10e4c38b0471b",
-  "publishConfig": {
-    "access": "public"
-  }
-}
