--- packages/modal/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "@fabric-ds/elements-modal",
-  "version": "0.0.3",
-  "description": "Custom elements for fabric",
-  "type": "module",
-  "exports": "./dist/index.js",
-  "module": "dist/fabric-modal.js",
-  "files": [
-    "dist/"
-  ],
-  "types": "dist/fabric-modal.d.ts",
-  "scripts": {
-    "build": "yarn build:js && yarn build:css",
-    "build:js": "esbuild src/fabric-modal.js --format=esm --outfile=dist/fabric-modal.js",
-    "build:css": "postcss -r dist/fabric-modal.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/fabric-ds/elements"
-  },
-  "license": "ISC",
-  "dependencies": {
-    "dom-focus-lock": "^1.0.4",
-    "lit-element": "^2.4.0",
-    "lit-html": "^1.3.0"
-  },
-  "gitHead": "a8154fa0605e4c7e3fa02f6058b10e4c38b0471b",
-  "publishConfig": {
-    "access": "public"
-  }
-}
