+++ packages/toast/index.js
@@ -3,7 +3,8 @@ import { FabricToast } from './component';
 
 initToasts();
 if (!customElements.get('f-toast')) {
-    customElements.define('f-toast', FabricToast);
+  customElements.define('f-toast', FabricToast);
 }
 
 export { toast, removeToast, updateToast } from './api';
+export { FabricToast };
