--- packages/toast/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@fabric-ds/toast",
-  "version": "0.0.7",
-  "type": "module",
-  "main": "./dist/index.js",
-  "types": "./dist/toast/src/index.d.ts",
-  "exports": "./dist/index.js",
-  "module": "./dist/index.js",
-  "files": [
-    "dist/"
-  ],
-  "scripts": {
-    "build": "yarn build:js",
-    "build:js": "microbundle -i src/index.js -o dist/index.js --no-pkg-main --format modern"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/fabric-ds/elements"
-  },
-  "license": "ISC",
-  "dependencies": {
-    "@chbphone55/classnames": "^2.0.0",
-    "@fabric-ds/component-classes": "^0.0.31",
-    "element-collapse": "^0.9.1"
-  },
-  "devDependencies": {
-    "microbundle": "^0.13.3",
-    "typescript": "^4.3.5"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "gitHead": "a8154fa0605e4c7e3fa02f6058b10e4c38b0471b"
-}
