+++ pages/includes/nav.html
@@ -7,10 +7,6 @@
         "open": true,
         "items": [
           {
-            "title": "Modal",
-            "href": "/pages/components/modal.html"
-          },
-          {
             "title": "Box",
             "href": "/pages/components/box.html"
           },
