+++ pages/includes/scripts.js
@@ -1,8 +1,1 @@
-import { FabricBox } from '../../packages/box/src/fabric-box.js';
-import { FabricBreadcrumbs } from '../../packages/breadcrumbs/src/fabric-breadcrumbs.js';
-import { FabricModal } from '../../packages/modal/src/fabric-modal.js';
-import '../../packages/toast/src/index.js';
-
-customElements.define('f-modal', FabricModal);
-customElements.define('f-box', FabricBox);
-customElements.define('f-breadcrumbs', FabricBreadcrumbs);
+import '../../index.js';
